# Minor fixes to InCodeGenerator