--- a/core/trino-main/src/main/java/io/trino/sql/gen/CompilerOperations.java
+++ b/core/trino-main/src/main/java/io/trino/sql/gen/CompilerOperations.java
@@ -19,7 +19,6 @@ import javax.annotation.Nullable;
 
 import java.util.List;
 import java.util.Optional;
-import java.util.Set;
 
 import static com.google.common.base.Preconditions.checkArgument;
 import static io.trino.spi.type.BooleanType.BOOLEAN;
@@ -58,11 +57,6 @@ public final class CompilerOperations
     public static boolean greaterThan(int left, int right)
     {
         return left > right;
-    }
-
-    public static boolean in(Object value, Set<?> set)
-    {
-        return set.contains(value);
     }
 
     public static boolean testMask(@Nullable Block masks, int index)