- Use enhanced instanceof
- Use enhanced switch
- Inline isInteger
- Use invokeVirtual on the set

In addition, to make sure that these changes do not cause performance
regressions, I also benchmarked `BIGINT` with `inListCount={5, 1000,
10000}`.

Before
```
Benchmark                           (hitRate)  (inListCount)  (type)  Mode  Cnt    Score     Error  Units
BenchmarkInCodeGenerator.benchmark        0.0              5  bigint  avgt   12   99.942 ±  26.115  us/op
BenchmarkInCodeGenerator.benchmark        0.0           1000  bigint  avgt   12  170.274 ± 108.237  us/op
BenchmarkInCodeGenerator.benchmark        0.0          10000  bigint  avgt   12  242.052 ±  83.163  us/op
BenchmarkInCodeGenerator.benchmark       0.05              5  bigint  avgt   12  149.536 ±  42.432  us/op
BenchmarkInCodeGenerator.benchmark       0.05           1000  bigint  avgt   12  112.260 ±  32.969  us/op
BenchmarkInCodeGenerator.benchmark       0.05          10000  bigint  avgt   12  207.537 ± 110.180  us/op
BenchmarkInCodeGenerator.benchmark       0.50              5  bigint  avgt   12  176.636 ±  55.588  us/op
BenchmarkInCodeGenerator.benchmark       0.50           1000  bigint  avgt   12  263.002 ± 133.450  us/op
BenchmarkInCodeGenerator.benchmark       0.50          10000  bigint  avgt   12  269.707 ± 128.761  us/op
BenchmarkInCodeGenerator.benchmark        1.0              5  bigint  avgt   12  102.594 ±  24.347  us/op
BenchmarkInCodeGenerator.benchmark        1.0           1000  bigint  avgt   12  129.428 ±  68.642  us/op
BenchmarkInCodeGenerator.benchmark        1.0          10000  bigint  avgt   12  182.918 ±  24.591  us/op
```

After
```
Benchmark                           (hitRate)  (inListCount)  (type)  Mode  Cnt    Score     Error  Units
BenchmarkInCodeGenerator.benchmark        0.0              5  bigint  avgt   12   79.320 ±   3.186  us/op
BenchmarkInCodeGenerator.benchmark        0.0           1000  bigint  avgt   12   82.954 ±  14.446  us/op
BenchmarkInCodeGenerator.benchmark        0.0          10000  bigint  avgt   12  184.912 ± 101.167  us/op
BenchmarkInCodeGenerator.benchmark       0.05              5  bigint  avgt   12  171.468 ±  73.439  us/op
BenchmarkInCodeGenerator.benchmark       0.05           1000  bigint  avgt   12  260.507 ±  50.637  us/op
BenchmarkInCodeGenerator.benchmark       0.05          10000  bigint  avgt   12  161.067 ±  70.446  us/op
BenchmarkInCodeGenerator.benchmark       0.50              5  bigint  avgt   12  142.667 ±  20.598  us/op
BenchmarkInCodeGenerator.benchmark       0.50           1000  bigint  avgt   12  206.942 ±  67.317  us/op
BenchmarkInCodeGenerator.benchmark       0.50          10000  bigint  avgt   12  182.434 ±  23.918  us/op
BenchmarkInCodeGenerator.benchmark        1.0              5  bigint  avgt   12   78.913 ±   5.840  us/op
BenchmarkInCodeGenerator.benchmark        1.0           1000  bigint  avgt   12   63.445 ±   1.933  us/op
BenchmarkInCodeGenerator.benchmark        1.0          10000  bigint  avgt   12  115.853 ±  49.151  us/op
```

## Release notes

(x) This is not user-visible or docs only and no release notes are required.